establishment delete borra la foto

+++ Backend/src/api/controllers/establishment.controller.js
@@ -144,15 +144,19 @@ const deleteEstablishment = async (req, res, next) => {
   const { id } = req.params;
 
   try {
+    const establishment = await Establishment.findById(id);
+    if (!establishment) {
+      return res.status(404).json({ error: "Establecimiento no encontrado" });
+    }
+    if (establishment.image) {
+      deleteImgCloudinary(establishment.image);
+    }
     const deletedEstablishment = await Establishment.findByIdAndDelete(id);
-
     if (deletedEstablishment) {
-      // Eliminar la referencia en los eventos
       await Event.updateMany(
         { establishment: id },
         { $pull: { establishment: id } },
       );
-
       return res
         .status(200)
         .json({ message: "Establecimiento eliminado exitosamente" });
